PDDA-474 Moving getCourtSites() out into an AbstractService layer for JudgeType
